refactor windows ipc into normal
