+++ src/ipc/connection.h
@@ -2,15 +2,26 @@
 #define IPC_CONNECTION_H
 
 #include <stddef.h>
+#ifdef __MSYS__
+#include <winsock2.h>
+#else
 #include <sys/un.h>
 #include <netinet/in.h>
+#include "socket.h"
+#endif
 
 struct connection {
-  int*                sock;
-  int*                msgsock;
+  SOCKET*             sock;
+#ifndef __MSYS__
+  SOCKET*             msgsock;
   struct sockaddr_un* server;
+#endif
   struct sockaddr_in* tcp_server;
+#ifdef __MSYS__
+  int msys_secret[4];
+#endif
 };
+
 
 int  connection_comparator(const struct connection* c1,
                            const struct connection* c2);
