+++ src/ipc/ipc.c
@@ -5,22 +5,28 @@
 #include "utils/logger.h"
 #include "utils/memory.h"
 #include "utils/oidc_error.h"
-#include "utils/printer.h"
 #include "utils/stringUtils.h"
 #ifdef __MSYS__
 #include "utils/registryConnector.h"
+#include "utils/file_io/file_io.h"
 #endif
 
+#ifdef __MSYS__
+#include <winsock2.h>
+#include <process.h>
+#else
 #include <arpa/inet.h>
-#include <stdarg.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
 #include <sys/ioctl.h>
 #include <sys/select.h>
 #include <sys/socket.h>
 #include <sys/un.h>
 #include <unistd.h>
+#endif
+#include <stdarg.h>
+#include <stdlib.h>
+#include <string.h>
+
+#ifndef __MSYS__
 
 oidc_error_t initConnectionWithoutPath(struct connection* con, int isServer,
                                        int tcp) {
@@ -62,21 +68,47 @@ oidc_error_t initConnectionWithPath(struct connection* con,
   return OIDC_SUCCESS;
 }
 
+#endif
+
 /**
  * @brief initializes a client unix domain or tcp socket
- * @param con, a pointer to the connection struct. The relevant fields will be
+ * @param con a pointer to the connection struct. The relevant fields will be
  * initialized.
- * @param env_var_name, the socket_path environment variable name
+ * @param remote determines whether its remote communication or not (remote communication currently not supported for Windows)
  */
 oidc_error_t ipc_client_init(struct connection* con, unsigned char remote) {
-  logger(DEBUG, "initializing client ipc");
+    logger(DEBUG, "initializing client ipc");
+#ifdef __MSYS__
+    if (remote) {
+        logger(DEBUG, "Remote connections are currently not supported by windows oidc-agent library");
+        return OIDC_EERROR;
+    }
+    con->tcp_server = secAlloc(sizeof(struct sockaddr_in));
+    con->sock = secAlloc(sizeof(SOCKET));
+
+    // Extract port and authorization string from MSYS socket file
+    int port;
+    sscanf(readFile(socketpath), "!<socket >%d s %X-%X-%X-%X", &port, &con->msys_secret[0], &con->msys_secret[1], &con->msys_secret[2], &con->msys_secret[3]);
+
+    con->msys_server->sin_port = htons(port);
+    con->msys_server->sin_addr.s_addr = inet_addr(SOCKET_LOOPBACK_ADDRESS);
+    con->msys_server->sin_family = AF_INET;
+
+    WSADATA wsa;
+    if (WSAStartup(MAKEWORD(2,2),&wsa) != 0)
+    {
+        return OIDC_ECONSOCK;
+    }
+
+    if((*(con->sock) = socket(AF_INET , SOCK_STREAM , 0 )) == INVALID_SOCKET)
+    {
+        return OIDC_ECRSOCK;
+    }
+    return OIDC_SUCCESS;
+#else
   const char* env_var_name =
       remote ? OIDC_REMOTE_SOCK_ENV_NAME : OIDC_SOCK_ENV_NAME;
-  #ifdef __MSYS__
-  char* path = getRegistryValue(env_var_name);
-  #else
-  char* path = oidc_strcopy(getenv(env_var_name));
-  #endif
+  char* path = getenv(env_var_name);
   if (path == NULL) {
     char* err = oidc_sprintf("Could not get the socket path from env var '%s'. "
                              "Have you set the env var?\n",
@@ -95,27 +127,37 @@ oidc_error_t ipc_client_init(struct connection* con, unsigned char remote) {
 
   if (remote) {
     logger(DEBUG, "Using TCP socket");
-    char*          ip         = strtok(path, ":");
+    char* tmpPath = oidc_strcopy(path);
+    char*          ip         = strtok(tmpPath, ":");
     char*          port_str   = strtok(NULL, ":");
     unsigned short port       = port_str == NULL ? 0 : strToUShort(port_str);
     con->tcp_server->sin_port = htons(port ?: 42424);
     con->tcp_server->sin_addr.s_addr =
         inet_addr(isValidIP(ip) ? ip : hostnameToIP(ip));
-    secFree(path);
+    secFree(tmpPath);
   } else {
     logger(DEBUG, "Using UNIX domain socket");
     strcpy(con->server->sun_path, path);
-    secFree(path);
   }
   return OIDC_SUCCESS;
+#endif
 }
 
 /**
  * @brief connects to a UNIX Domain or TCP socket
- * @param con, the connection struct
- * @return the socket or @c OIDC_ECONSOCK on failure
+ * @param con the connection struct
+ * @return @c OIDC_SUCCESS or @c OIDC_ECONSOCK on failure
  */
-int ipc_connect(struct connection con) {
+oidc_error_t ipc_connect(struct connection con) {
+#ifdef __MSYS__
+    if (connect(*(con.sock), (struct sockaddr *)con.tcp_server , sizeof(struct sockaddr_in)) < 0) {
+        closesocket(*(con.sock));
+        WSACleanup();
+        oidc_errno = OIDC_ECONSOCK;
+        return oidc_errno;
+    }
+    return OIDC_SUCCESS;
+#else
   struct sockaddr* server      = (struct sockaddr*)con.server;
   size_t           server_size = sizeof(struct sockaddr_un);
   if (con.server->sun_path[0] == '\0') {
@@ -130,10 +172,53 @@ int ipc_connect(struct connection con) {
     close(*(con.sock));
     logger(ERROR, "connecting stream socket: %m");
     oidc_errno = OIDC_ECONSOCK;
-    return OIDC_ECONSOCK;
+    return oidc_errno;
   }
-  return *(con.sock);
+  return OIDC_SUCCESS;
+#endif
 }
+
+#ifdef __MSYS__
+/**
+ * @brief authorizes against msys emulated socket
+ * @param con, the connection struct
+ * @return @c OIDC_SUCCESS or @c OIDC_EMSYSAUTH on failure
+ */
+oidc_error_t ipc_msys_authorize(struct connection con) {
+    char *ptr = (char *) con.msys_secret;
+    int written_bytes = send(*(con.sock), ptr , sizeof con.msys_secret , 0);
+    if (written_bytes != sizeof con.msys_secret) {
+        oidc_errno = OIDC_EMSYSAUTH;
+        return oidc_errno;
+    }
+
+    int out[4] = { 0, 0, 0, 0};
+    ptr = (char *) out;
+    int recv_size = recv(*(con.sock), ptr , sizeof out , 0);
+    if (recv_size != sizeof out) {
+        oidc_errno = OIDC_EMSYSAUTH;
+        return oidc_errno;
+    }
+
+    int messageCred[3] = {_getpid(), -1, -1};
+    ptr = (char *) messageCred;
+    written_bytes = send(*(con.sock) , ptr , sizeof messageCred , 0);
+    if (written_bytes != sizeof messageCred) {
+        oidc_errno = OIDC_EMSYSAUTH;
+        return oidc_errno;
+    }
+
+    int outCred[3] = {0, 0, 0};
+    ptr = (char *) outCred;
+    recv_size = recv(*(con.sock) , ptr, sizeof outCred , 0);
+    if (recv_size != sizeof outCred) {
+        oidc_errno = OIDC_EMSYSAUTH;
+        return oidc_errno;
+    }
+
+    return OIDC_SUCCESS;
+}
+#endif
 
 /**
  * @brief reads from a socket until a timeout is reached
@@ -142,8 +227,57 @@ int ipc_connect(struct connection con) {
  * @return a pointer to the readed content. Has to be freed after usage. If
  * @c NULL is returned, it's most likely that the other party disconnected.
  */
-char* ipc_read(const int _sock) { return ipc_readWithTimeout(_sock, 0); }
+char* ipc_read(const SOCKET _sock) {
+#ifdef __MSYS__
+    logger(DEBUG, "ipc reading from socket %d\n", _sock);
 
+    u_long len = 0;
+    int rv;
+    fd_set set;
+    FD_ZERO(&set);
+    FD_SET(_sock, &set);
+
+    rv = select(_sock + 1, &set, NULL, NULL, 0);
+    if (rv == -1) {
+        logger(ALERT, "error select in %s: %m", __func__);
+        oidc_errno = OIDC_ESELECT;
+        return NULL;
+    }
+    if (rv == 0) {
+        oidc_errno = OIDC_ETIMEOUT;
+        return NULL;
+    }
+    if (ioctlsocket(_sock, FIONREAD, &len) != 0) {
+        logger(ERROR, "ioctl: %m");
+        oidc_errno = OIDC_EIOCTL;
+        return NULL;
+    }
+    if (len <= 0) {
+        logger(DEBUG, "Client disconnected");
+        oidc_errno = OIDC_EIPCDIS;
+        return NULL;
+    }
+    char* buf = secAlloc(sizeof(char) * (len + 1));
+    logger(DEBUG, "ipc want to read %d bytes", len);
+    int read_bytes = 0;
+    while (read_bytes < (int) len) {
+        int read_ret = recv(_sock, buf + read_bytes, len - read_bytes, 0);
+        if (read_ret < 0) {
+            oidc_setErrnoError();
+            secFree(buf);
+            return NULL;
+        }
+        read_bytes += read_ret;
+        logger(DEBUG, "ipc did read %d bytes in total", read_bytes);
+    }
+    logger(DEBUG, "ipc read '%s'", buf);
+    return buf;
+#else
+    return ipc_readWithTimeout(_sock, 0);
+#endif
+}
+
+#ifndef __MSYS__
 struct timeval* initTimeout(time_t death) {
   if (death == 0) {
     oidc_errno = OIDC_SUCCESS;
@@ -221,6 +355,7 @@ char* ipc_readWithTimeout(const int _sock, time_t death) {
   logger(DEBUG, "ipc read '%s'", buf);
   return buf;
 }
+#endif
 
 /**
  * @brief writes a message to a socket
@@ -228,7 +363,7 @@ char* ipc_readWithTimeout(const int _sock, time_t death) {
  * @param msg the msg to be written
  * @return @c 0 on success; on failure an error code is returned
  */
-oidc_error_t ipc_write(int _sock, const char* fmt, ...) {
+oidc_error_t ipc_write(SOCKET _sock, const char* fmt, ...) {
   va_list args;
   va_start(args, fmt);
   oidc_error_t ret = ipc_vwrite(_sock, fmt, args);
@@ -236,12 +371,16 @@ oidc_error_t ipc_write(int _sock, const char* fmt, ...) {
   return ret;
 }
 
-oidc_error_t ipc_vwrite(int _sock, const char* fmt, va_list args) {
+oidc_error_t ipc_vwrite(SOCKET _sock, const char* fmt, va_list args) {
   char* msg = oidc_vsprintf(fmt, args);
   if (msg == NULL) {
     return oidc_errno;
   }
+#ifdef __MSYS__
+  int msg_len = strlen(msg);
+#else
   size_t msg_len = strlen(msg);
+#endif
   if (msg_len == 0) {  // Don't send an empty message. This will be read as
                        // client disconnected
     msg_len = 1;
@@ -250,21 +389,29 @@ oidc_error_t ipc_vwrite(int _sock, const char* fmt, va_list args) {
   }
   logger(DEBUG, "ipc writing %lu bytes to socket %d", msg_len, _sock);
   logger(DEBUG, "ipc write message '%s'", msg);
+#ifdef __MSYS__
+  int written_bytes = write(_sock, msg, msg_len);
+#else
   ssize_t written_bytes = write(_sock, msg, msg_len);
+#endif
   secFree(msg);
   if (written_bytes < 0) {
     logger(ALERT, "writing on stream socket: %m");
     oidc_errno = OIDC_EWRITE;
     return oidc_errno;
   }
+#ifdef __MSYS__
+  if (written_bytes < msg_len) {
+#else
   if ((size_t)written_bytes < msg_len) {
+#endif
     oidc_errno = OIDC_EMSGSIZE;
     return oidc_errno;
   }
   return OIDC_SUCCESS;
 }
 
-oidc_error_t ipc_writeOidcErrno(int sock) {
+oidc_error_t ipc_writeOidcErrno(SOCKET sock) {
   return ipc_write(sock, RESPONSE_ERROR, oidc_serror());
 }
 
@@ -272,7 +419,14 @@ oidc_error_t ipc_writeOidcErrno(int sock) {
  * @brief closes a FD
  * @param _sock the FD to be closed
  */
-int ipc_close(int _sock) { return close(_sock); }
+int ipc_close(int _sock) {
+#ifdef __MSYS__
+    WSACleanup();
+    return closesocket(_sock);
+#else
+    return close(_sock);
+#endif
+}
 
 /**
  * @brief closes an ipc connection
@@ -284,40 +438,24 @@ oidc_error_t ipc_closeConnection(struct connection* con) {
   if (con->sock != NULL) {
     ipc_close(*(con->sock));
   }
+#ifndef __MSYS__
   if (con->msgsock != NULL) {
     ipc_close(*(con->msgsock));
   }
   secFree(con->server);
-  con->server = NULL;
-  secFree(con->tcp_server);
-  con->tcp_server = NULL;
-  secFree(con->sock);
-  con->sock = NULL;
   secFree(con->msgsock);
-  con->msgsock = NULL;
+#endif
+  secFree(con->sock);
+  secFree(con->tcp_server);
   return OIDC_SUCCESS;
 }
 
-/**
- * @brief closes an ipc connection and removes the socket
- * @param con, a pointer to the connection struct
- * @return @c OIDC_SUCCESS on success
- */
-oidc_error_t ipc_closeAndUnlinkConnection(struct connection* con) {
-  if (con->server->sun_path != NULL) {
-    logger(DEBUG, "Unlinking %s", con->server->sun_path);
-    unlink(con->server->sun_path);
-  }
-  ipc_closeConnection(con);
-  return OIDC_SUCCESS;
-}
-
-char* ipc_vcommunicateWithSock(int sock, const char* fmt, va_list args) {
+char* ipc_vcommunicateWithSock(SOCKET sock, const char* fmt, va_list args) {
   ipc_vwrite(sock, fmt, args);
   return ipc_read(sock);
 }
 
-char* ipc_communicateWithSock(int sock, const char* fmt, ...) {
+char* ipc_communicateWithSock(SOCKET sock, const char* fmt, ...) {
   va_list args;
   va_start(args, fmt);
   char* ret = ipc_vcommunicateWithSock(sock, fmt, args);
