+++ src/ipc/socket.h
@@ -1,0 +1,8 @@
+#ifndef OIDC_AGENT_SOCKET_H
+#define OIDC_AGENT_SOCKET_H
+
+#ifndef SOCKET
+#define SOCKET int
+#endif
+
+#endif //OIDC_AGENT_SOCKET_H
