--- src/ipc/windows/cryptCommunicator.h
@@ -1,8 +1,0 @@
-#ifndef CRYPT_COMMUNICATOR_H
-#define CRYPT_COMMUNICATOR_H
-
-#include <stdarg.h>
-
-char* ipc_vcryptCommunicate(unsigned char, const char*, va_list);
-
-#endif  // CRYPT_COMMUNICATOR_H
